Move unescape helper to markdown-it
