+++ packages/core/src/utils/index.js
@@ -9,14 +9,6 @@ const {
   markdownFileExts,
 } = require('../constants');
 
-const htmlUnescapedMapping = {
-  '&amp;': '&',
-  '&lt;': '<',
-  '&gt;': '>',
-  '&quot;': '"',
-  '&#x27;': '\'',
-};
-
 module.exports = {
   getCurrentDirectoryBase() {
     return path.basename(process.cwd());
@@ -112,12 +104,5 @@ module.exports = {
     }
 
     return text.join('').trim();
-  },
-  unescapeHtml(str) {
-    let unescaped = str;
-    Object.entries(htmlUnescapedMapping).forEach(([key, value]) => {
-      unescaped = unescaped.split(key).join(value);
-    });
-    return unescaped;
   },
 };
